Corrections after testing
Corrections were made in the repository interfaces related to hibernate due to an incorrect handling of session closing.
Added some missing fields in the jsons and also added an endpoint to see the books currently borrowed by the user.
Added id field to borrowed record due to uniqueness issues.
finally, several refactors were implemented to improve the readability of the code and to have simpler classes.